pyproject.toml
[project]
name = "repiko"
version = "0.1.0"
description = "一个基于OneBot的复读机"
authors = [
    { name = "liggest" },
]
dependencies = [
    "lsparser>=0.1.7",
    "fastapi[all]>=0.86.0",
    "beautifulsoup4>=4.11.1",
    "pillow>=9.3.0",
    "lxml>=4.9.1",
    "requests>=2.28.1",
    "deepl>=1.12.0",
    "loguru>=0.6.0",
    "ruamel.yaml>=0.17.21",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
[project.optional-dependencies]
svg = [
    "svglib>=1.4.1",
    "pycairo>=1.23.0",
    "cairocffi>=1.4.0",
    "rlpycairo>=0.1.0",
]
chat = [
    "revChatGPT>=3.1.3.1",
]

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
app = "python bot.py {args:--host=0.0.0.0 --port=5701 --reload --reload-dir repiko}"
# app = "uvicorn bot:app --reload --reload-dir repiko --host=0.0.0.0 {args:--port=8080}"
# wsapp = "python -m repiko.core.websocket"
# watch = "watchfiles 'pdm run app' repiko"
